- CAPACITY_EVENT type: ticketed/carpool events with seat limits
- TRAVEL category: Google Maps route with proportional km pricing
- Route stops: host adds stops via Places autocomplete + Directions API
- Guest flow: browse events, join with GPS-nearest stop, UPI payment
- DB migrations V10-V15: capacity fields, route stops, booking segments
- EventResponse merged: includes cash/UPI amounts + routeStops/capacity
- EventDetailScreen: shows capacity stats when isCapacity, gift stats otherwise
- Accept Gift button shown only for GIFT_COLLECTION events

+++ chanlo/src/main/java/com/mysteriousmonkeys/chanlo/dto/GuestUpiPaymentRequest.java
@@ -27,5 +27,5 @@ public record GuestUpiPaymentRequest(
 
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be greater than 0")
-    Long amount
+    Double amount
 ) {}
